bux fixes with views

+++ test_project/views.py
@@ -1,7 +1,7 @@
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from .models import Product, Cart, CartItem, Order, OrderItem, Category, Review, UserProfile
-
+from django.db.models import Sum
 
 def product_list(request):
     products = Product.objects.all()
@@ -46,7 +46,7 @@ def add_to_cart(request, product_id):
 def cart_detail(request):
     cart, _ = Cart.objects.get_or_create(user=request.user)
     cart_items = list(cart.cartitem_set.all().values())
-    total_price = str(cart.total_price)  # Преобразуем Decimal в строку
+    total_price = cart.cartitem_set.aggregate(total_price=Sum('product__price'))['total_price'] or 0
     data = {"cart": {"items": cart_items, "total_price": total_price}}
     return JsonResponse(data)
 
@@ -55,11 +55,15 @@ def cart_detail(request):
 
 def create_order(request):
     cart, _ = Cart.objects.get_or_create(user=request.user)
-    order = Order.objects.create(user=request.user, total_price=cart.total_price, status='pending')
-    for cart_item in cart.cartitem_set.all():
-        OrderItem.objects.create(order=order, product=cart_item.product, quantity=cart_item.quantity)
-    cart.delete()
-    return JsonResponse({"status": "success"})
+    total_price = cart.total_price if hasattr(cart, 'total_price') else None
+    if total_price is not None:
+        order = Order.objects.create(user=request.user, total_price=total_price, status='pending')
+        for cart_item in cart.cartitem_set.all():
+            OrderItem.objects.create(order=order, product=cart_item.product, quantity=cart_item.quantity)
+        cart.delete()
+        return JsonResponse({"status": "success"})
+    else:
+        return JsonResponse({"status": "error", "message": "Total price is not available"})
 
 
 def order_confirmation(request):
